Fix AUXR and PPG channels not streaming under bgapi
It appears that muse (self.select_preset) was no longer being called when
the streaming started. This prevented the select preset command from being
sent to the headset, which means that it would never sample data on anything
except for the default channels.

At this time, it is not enough to subscribe to the GATT attributes of the
desired data, the correct preset number must also be sent in order to
instruct the headset which channels are to be sampled.

I have added a new command line argument '-P' '--preset', which allows an
integer to be passed for selecting the preset. For reference::

    20 - 5-Channel EEG channel streaming
    21 - 4-Channel EEG channel streaming
    22 - 4-Channel EEG channel streaming without accel/gyro

    50 - Same as 20, but includes PPG data
    51 - Same as 21, but includes PPG data

I have not yet tested this with a backend other than BGAPI, so I'm not
sure if Bluemuse will send the preset information either.